pyproject.toml
[project]
name = "port-report"
version = "0.1.0"
description = "A project where u enter your ip and it uses agents and all "
readme = "README.md"
authors = [
    { name = "Akashthakur0326", email = "akashthakur0326@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "streamlit",
    "requests",
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0"
]

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"
